fix(gsd): preserve inspected empty subagent inputs
